RHICOMPL-449 - Fix floating point thresholds (#268)

+++ src/SmartComponents/EditPolicy/EditPolicy.js
@@ -78,7 +78,7 @@ export class EditPolicy extends Component {
                         <UpdateProfileButton
                             key='confirm'
                             policyId={policyId}
-                            threshold={ parseInt(complianceThreshold || previousThreshold) }
+                            threshold={ parseFloat(complianceThreshold || previousThreshold) }
                             businessObjective={businessObjective}
                             editPolicyBusinessObjective={editPolicyBusinessObjective}
                             onClick={this.handleModalToggle}
